requirements.txt
discord.py==2.0.1
lavaplayer==1.0.10a0
PyNaCl==1.5.0